+++ bugsnag-android-core/src/main/java/com/bugsnag/android/User.kt
@@ -1,5 +1,6 @@
 package com.bugsnag.android
 
+import android.util.JsonReader
 import java.io.IOException
 
 /**
@@ -25,10 +26,39 @@ class User @JvmOverloads internal constructor(
     @Throws(IOException::class)
     override fun toStream(writer: JsonStream) {
         writer.beginObject()
-        writer.name("id").value(id)
-        writer.name("email").value(email)
-        writer.name("name").value(name)
+        writer.name(KEY_ID).value(id)
+        writer.name(KEY_EMAIL).value(email)
+        writer.name(KEY_NAME).value(name)
         writer.endObject()
+    }
+
+    internal companion object: JsonReadable<User> {
+        private const val KEY_ID = "id"
+        private const val KEY_NAME = "name"
+        private const val KEY_EMAIL = "email"
+
+        override fun fromReader(reader: JsonReader): User {
+            var user: User
+            with(reader) {
+                beginObject()
+                var id: String? = null
+                var email: String? = null
+                var name: String? = null
+
+                while (hasNext()) {
+                    val key = nextName()
+                    val value = nextString()
+                    when (key) {
+                        KEY_ID -> id = value
+                        KEY_EMAIL -> email = value
+                        KEY_NAME -> name = value
+                    }
+                }
+                user = User(id, email, name)
+                endObject()
+            }
+            return user
+        }
     }
 
     override fun equals(other: Any?): Boolean {